interpreter: improve plug pattern matching (#181)
## Summary

Improve meta-language interpreter's plug pattern matching, making holes
following repetitions in lists work as expected and preparing for
further improvements.

## Details

Instead of treating hole patterns as matching everything and then
handling the match between the captured term and plugged-with pattern
once the total match has been successful, the plugged-with pattern is
used in place of the hole.

Passing values "downwards" during pattern matching is not supported, so
instead of passing the plugged-with pattern towards the hole matcher,
a `matches` encountering a hole returns the special `Match` instance
(using the `hole` kind) that contains the current continuation to the
closest handler, which then handles the hole match appropriately.

To support multi-result patterns in a more generally applicable way,
successful matches now store an optional continuation representing a
not-yet-tried match thread.

Together this makes the following pattern usages work (without an
external match driver):
1. holes following repetitions in lists
2. using a context name as a pattern in contexts
3. plug patterns part of contexts
4. holes appearing in repetitions

Only number 1 actually affects the meta language, as 2 was previously
implemented via an external match driver and both 2 and 3 are either
forbidden or already handled/inlined by the type checker.

Finally, some tests are added directly for the pattern matcher to make
sure the changes work.