nanopass: use a different approach to terminal storage
Instead of requiring the programmer to pass the storage instance to
every pass (and carry it around to everywhere the AST is to be used),
a ref to the storage is now stored within the AST itself.

Storage types not named `Literals` and using pre-existing storage
object instance for newly-created AST are not supported yet.